Add delta connector
