+++ crates/data_components/src/lib.rs
@@ -31,6 +31,8 @@ pub mod databricks_spark;
 pub mod debezium;
 #[cfg(feature = "debezium")]
 pub mod debezium_kafka;
+#[cfg(feature = "databricks")]
+pub mod delta;
 #[cfg(feature = "duckdb")]
 pub mod duckdb;
 pub mod flight;
