+++ crates/runtime/src/dataconnector.rs
@@ -59,6 +59,8 @@ pub mod clickhouse;
 pub mod databricks;
 #[cfg(feature = "debezium")]
 pub mod debezium;
+#[cfg(feature = "databricks")]
+pub mod delta;
 #[cfg(feature = "dremio")]
 pub mod dremio;
 #[cfg(feature = "duckdb")]
@@ -246,6 +248,8 @@ pub async fn register_all() {
     register_connector_factory("localhost", localhost::LocalhostConnector::create).await;
     #[cfg(feature = "databricks")]
     register_connector_factory("databricks", databricks::Databricks::create).await;
+    #[cfg(feature = "databricks")]
+    register_connector_factory("delta", delta::Delta::create).await;
     #[cfg(feature = "dremio")]
     register_connector_factory("dremio", dremio::Dremio::create).await;
     register_connector_factory("file", file::File::create).await;
